Split dataset selection into a separate Search Catalog app
Extract the prefilled-URL table + URL input + featureType-detect/redirect from
TSP/main.py into a new metviz/catalog app, served as /Catalog ("Search
Catalog"). The Panel server now serves three apps: TSP (timeSeries/profile/
timeSeriesProfile), TRJ (trajectory), and Catalog (dataset selection). TSP with
no `url` now shows a notice linking to /Catalog instead of the table.

This is the seed for an OGC CSW catalogue search that will replace/augment the
static example table. Wires /Catalog into entrypoint.sh and the docker-compose
mounts; ncmet.yml (gitignored) updated on disk to match.
