+++ ruff.toml
@@ -18,3 +18,4 @@ ignore = [
 "ncapp/app/main.py" = ["E402"]
 "metviz/TSP/main.py" = ["E402"]
 "metviz/trj/main.py" = ["E402"]
+"metviz/catalog/main.py" = ["E402"]
